+++ apollo-portal/src/main/resources/static/scripts/directive/item-modal-directive.js
@@ -39,11 +39,20 @@ function itemModalDirective($translate, toastr, $sce, AppUtil, EventManager, Con
             scope.doItem = doItem;
             scope.collectSelectedClusters = collectSelectedClusters;
             scope.showHiddenChars = showHiddenChars;
+            scope.detectJSON = detectJSON;
+            scope.check = check;
 
             $('#itemModal').on('show.bs.modal', function (e) {
                 scope.showHiddenCharsContext = false;
                 scope.hiddenCharCounter = 0;
                 scope.valueWithHiddenChars = $sce.trustAsHtml('');
+                scope.showCheckJSONHint = false;
+                scope.showJSONDetectContext = false;
+                scope.jsonDetectResult = $sce.trustAsHtml('');
+            });
+
+            $('#itemModal').on('shown.bs.modal', function (e) {
+                scope.$apply("check()");
             });
 
             $("#valueEditor").textareafullscreen();
@@ -152,6 +161,40 @@ function itemModalDirective($translate, toastr, $sce, AppUtil, EventManager, Con
 
             function collectSelectedClusters(data) {
                 selectedClusters = data;
+            }
+
+            function check(){
+                var value = scope.item.value;
+                if (!value || value.length < 2) {
+                    scope.showCheckJSONHint = false;
+                    scope.showJSONDetectContext = false;
+                    return;
+                }
+                //this way may more effective than regex
+                var signal = value.charAt(0)+value.charAt(value.length-1);
+                if(signal === "{}" || signal === "[]"){
+                    scope.showCheckJSONHint = true;
+                }else{
+                    scope.showCheckJSONHint = false;
+                    scope.showJSONDetectContext = false;
+                }
+            }
+
+            function detectJSON() {
+                var value = scope.item.value;
+                if (!value) {
+                    scope.showJSONDetectContext = false;
+                    return;
+                }
+                var res = "";
+                try {
+                    JSON.parse(value);
+                    res = $translate.instant('Component.ConfigItem.ValidItemJSONValue')
+                } catch(e) {
+                    res = $translate.instant('Component.ConfigItem.InvalidItemJSONValue')
+                }
+                scope.showJSONDetectContext = true;
+                scope.jsonDetectResult = $sce.trustAsHtml(res);
             }
 
             function showHiddenChars() {